Move DiscardEntityChangesPanel to dedicated file
